+++ src/sdk/constants/addresses.ts
@@ -205,3 +205,34 @@ export const WNATIVE_ADDRESS: AddressMap = {
 export const ZAPPER_ADDRESS: AddressMap = {
   [ChainId.ETHEREUM]: '0xcff6eF0B9916682B37D80c19cFF8949bc1886bC2',
 }
+
+export const LAMBDA_URL = 'https://9epjsvomc4.execute-api.us-east-1.amazonaws.com/dev'
+export const SOCKET_URL = 'wss://hfimt374ge.execute-api.us-east-1.amazonaws.com/dev'
+
+export const STOP_LIMIT_ORDER_ADDRESS: AddressMap = {
+//   [ChainId.AVALANCHE]: '0xf6f9c9DB78AF5791A296c4bF34d59E0236E990E0',
+  [ChainId.FANTOM]: '0x6AaC28444cB227EC5D5cebe696882DeCcD50ee83', // FEB22
+  [ChainId.ETHEREUM]: '0x104734Ce12567421aC7B562e03Be2D75522cd112',
+}
+
+export const DEFAULT_RECEIVER_ADDRESS: AddressMap = {
+//   [ChainId.AVALANCHE]: '0x042c99C84b00f11A08a07AA9752E083261083A57',
+  [ChainId.FANTOM]: '0x612D3c387c2A483084D68061c753Ce1AD4e88bb6', // FEB22
+  [ChainId.ETHEREUM]: '0xf4943f2dEc7E4914067CdF4120E8A322bc8f5a36', // SSLO
+}
+
+export const ADVANCED_RECEIVER_ADDRESS: AddressMap = {
+//   [ChainId.AVALANCHE]: '0x50995361A1104B2E34d81771B2cf19BA55051C7c',
+  [ChainId.FANTOM]: '0xd6AF3AAe2Aef4f1Acff9dD66f542ea863fBe9ae7', // FEB22
+  [ChainId.ETHEREUM]: '0xA32e906C31093aDbe581B913e549f70fD2fD7969', // SSLO2
+}
+
+export const LIMIT_ORDER_HELPER_ADDRESS: AddressMap = {
+//   [ChainId.AVALANCHE]: '0x889ec9e19C1598358899fCA4879011686c3d4045',
+  [ChainId.FANTOM]: '0xd63E7D4eB9aB59bf85975c7100a5D92919C4E7E5', // FEB22
+  [ChainId.ETHEREUM]: '0x75a5263bddD871E94188611f3563aabc833Cc005',
+}
+
+export const ROUND_UP_RECEIVER_ADDRESS: AddressMap = {
+//   [ChainId.MATIC]: '0x1C9B033F8C46C08EbE67F15924F5B9E97e36E0a7'
+}
