+++ src/sdk/entities/index.ts
@@ -17,4 +17,4 @@ export * from './Token'
 export * from './TridentTrade'
 export * from './Trade'
 export * from './Native'
-export * from '../limit-order/eip712'
+export * from './LimitOrder'
