+++ src/sdk/enums/index.ts
@@ -1,6 +1,7 @@
 export { ChainId } from './ChainId'
 export { ChainKey } from './ChainKey'
 export { Fee } from './Fee'
+export { OrderStatus } from './OrderStatus'
 export { PoolState } from './PoolState'
 export { Rounding } from './Rounding'
 export { RouteType } from './RouteType'
