+++ src/sdk/index.ts
@@ -14,4 +14,3 @@ export * from './router'
 export * from '../utils'
 export * from '../types'
 export * from '../utils'
-export * from './limit-order'
