+++ src/sdk/interfaces/index.ts
@@ -7,3 +7,4 @@ export type { ExactInputParams } from './ExactInputParams'
 export type { ExactInputSingleParams } from './ExactInputSingleParams'
 export type { TridentRoute } from './TridentRoute'
 export type { PercentagePath } from './PercentagePath'
+export * from './eip'
