--- src/sdk/limit-order/index.ts
@@ -1,5 +1,0 @@
-export * from './constants'
-export * from './eip712'
-export * from './enums'
-export * from './limitorder'
-export * from './types'
